---Refactoring Packaging--- Sorting every class and component into packages that properly represent the structure of the code
