+++ src/components/webDesign/navBar/Navbar.js
@@ -1,6 +1,6 @@
 import React, {useState, useEffect} from 'react'
 import { Link } from 'react-router-dom';
-import { Button } from './Button';
+import { Button } from '../button/Button';
 import './Navbar.css';
 
 /**
